Fixed issues for mamda's forgiving type casting
For books to work, it needed to be possible to use getF64 to
retrieve a value which is in fact a U32 in most modern messages.

Will add a MAMA unit test to test for this as well because unit
tests really should have caught this it was a nightmare to trace
down.